# models/trail.py
from . import db
from .location import Location

class Trail(db.Model):
    __tablename__ = 'Trail'
    __table_args__ = {'schema': 'Referral'}

    TrailID = db.Column(db.Integer, primary_key=True)
    TrailName = db.Column(db.String(100), nullable=False)
    TrailRating = db.Column(db.Numeric(5, 3), nullable=True)
    TrailDifficulty = db.Column(db.String(10), nullable=False)
    TrailDistance = db.Column(db.Numeric(5, 2), nullable=False)
    TrailEstTime = db.Column(db.String(10), nullable=False)
    TrailRouteType = db.Column(db.String(10), nullable=False)
    TrailDescription = db.Column(db.Text, nullable=False)
    LocationID = db.Column(db.Integer, db.ForeignKey('Referral.Location.LocationID'), nullable=False)


    location = db.relationship('Location', backref=db.backref('trails', lazy=True))